Switch to x86_64 and NASM Assembly
This is a big one lmao

So the main thing here is that I switched the kernel over from i386 to x86_64. There might be some comments here or there saying the latter but we're definitely in long mode lads. I also switched from GAS to NASM since NASM's b e t t e r and I needed to do a bunch of assembly coding for this and I didn't wanna do it in GAS.

This was a long day im going to bed now